+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 - "payment" account type for Pix related resources
 - missing parameters to Boleto, BrcodePayment, DictKey, Event, Transfer and Workspace resources
 - workspace/update to allow parameter updatesa
+- event.attempt sub-resource to allow retrieval of information on failed webhook event delivery attempts
 
 ## [2.3.4] - 2021-03-04
 ### Added